== Compositor == * Displace Node
Displaces an input image's pixels based on an input vector mask. This can be
useful for a lot of things, like hot air distortion, quick-and-dirty compo
refraction, compositing live footage behind refracting objects, and more!

The amount of displacement in the X and Y directions is determined by:
* The value of the mask's channels
  - (red) channel 1's value determines displacement along the positive or
  negative X axis
  - (green) channel 2's value determines displacement along the positive or
  negative Y axis

If both the channel's values are equal (i.e. a greyscale image) the input
image will be displaced equally in both X and Y directions, also according to:
* The X scale and Y scale buttons
  - These act as multipliers to increase or decrease the strength of the
  displacement along their respective axes. They need to be set to non-zero
  values for the node to have any effect.

Because of this, you can use the displace node in two ways, with a greyscale
mask(easy to paint, or take from a procedural texture), or with a vector
channel or RGB image, such as a normal pass, which will displace the pixels
based on the normal direction.

A quick practical example:
http://mke3.net/blender/etc/displace-desert-h264.mov
http://mke3.net/blender/etc/displace-desert.blend.zip

And some techie examples:

Using a greyscale mask
http://mke3.net/blender/etc/displace-bw-h264.mov
http://mke3.net/blender/etc/displace-bw.png

Using a vector mask
http://mke3.net/blender/etc/displace-vec-h264.mov
http://mke3.net/blender/etc/displace-vec.png


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@9545 954f8c5b-7b00-dc11-b283-0030488c597c