Bugfix #5750
<sob>the 'time cursor' for bake-render has to go... it is being called
inside of a thread (even whilst thread is mutex locked), but that's not
supported in X11. It might be even instable in OSX/Windows even...

Only way to bring it back is to have the main loop (not in thread) update
the time cursor like each second. Would still mean to add a counter var...
will think it over. First want to have confirmed this is stable.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@9839 954f8c5b-7b00-dc11-b283-0030488c597c