Bugfix #5833:
When rotating bones around the 3d-cursor in posemode with a rotated
armature, rotation was around strange points other than the cursor.

This bug has been around for quite a few releases now. Somehow, the
maths used to convert the world/global space locations to local locations
only worked on things in editmode, but not bones.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@10011 954f8c5b-7b00-dc11-b283-0030488c597c