=== Transform Snap ===
Snapping to other meshes from edit mode.

This commit adds the capacity to snap to vertice of other selected meshes. (This is for sanity reasons, big scenes would kill the speed)
So, if you want to snap to another mesh, just add it to the selection and you're done.

NOTE: Priority is given to vertice inside the current mesh (if they overlap with outside verts)


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@10169 954f8c5b-7b00-dc11-b283-0030488c597c