Compositor/assorted nodes fixes:
* Auto-create compbufs for first socket of alphaover and set alpha nodes.
This allows you to eg. plug something into the second socket of an
alphaover node, and choose a solid colour in the first empty socket for
the image to be superimposed over.

Previously I had to create a bunch of extra nodes to (for example) mix
100% black over my render, just to get a black compbuf of the right size
that I could plug in. Not nice.

The Mix node already works this way, and these two should have, but
didn't.

* Allow the 'Fac' value to be used on RGB curves when there is no input
image. This lets you easily fade the changes in and out to check it
against the original, or to tone down the effect of the colour correction.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@10215 954f8c5b-7b00-dc11-b283-0030488c597c