Fix for bug #7406 'Compositor Crash'
Previously when deleting a scene, no checks were made to see if 
any render layer nodes in other scenes were referencing them,
leaving them pointing to bad memory. Now, this checks all other
scenes' nodetrees and sets pointers to the to-be-deleted scene to
NULL, which gets taken care of elsewhere.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@11580 954f8c5b-7b00-dc11-b283-0030488c597c