BugFix #7115
The "Strands" settings is a BlockBut and never return B_MATPRV like the others buttons, because of this,
never call BIF_preview_changed.

It's posible two fix for this:
	1) make all the buttons into the "Strands" panel return B_MATPRV (this modify the "work mode" of the panel)
	2) Add a call to BIF_preview_changed to the end of strands_menu function (as do_matbuts).

The problem with the first options, is that the panel "close" when you press/release some of the buttons, so
to preserve the current "work mode" of the panel, i choose the second option.



git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@11839 954f8c5b-7b00-dc11-b283-0030488c597c