A user submitted a BVH file that took a long time to import (I didnt end up finishing since it was so slow) this is mainly because adding pose keyframes recalculates every handle so importing became increasingly slow.
added a 'fast' argument to insertkey that python api's insertPoseKey can make use of since it alredy accepts a 'fast' option.

The ~4450 frame, 31 bone BVH imports in ~108sec now

Seperated editmode switch statement in space.c's event handling,
if editmode is disabled, or the images is a render or composite, UV editing operations are ignored.
In previous releases it has given an annoying warning if selecting or scaling is attempted when out of UV/Face mode.



git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@12078 954f8c5b-7b00-dc11-b283-0030488c597c