== Armature Layer/Move Bone to Layer Popups ==
This commit adds a few quick tools for riggers. In Editmode for Armatures, I've added the popups that show up in PoseMode when Shift-MKEY and MKEY are used. This should speed up the workflow a bit, by requiring less trips between the 3d-view and the buttons panel, as well as providing a 'batch' move-bones-to-layer functionality.

Usage Notes:
* MKEY - move selected bones to layer(s)
* SHIFT-MKEY - change the currently visible armature layers
* CTRL-MKEY - mirror selected bones 

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@12195 954f8c5b-7b00-dc11-b283-0030488c597c