fix for crash's in file selector. - on unix BLI_diskfree was only using 100 chars for the dir name, and not checking if the name given was longer, increased to FILE_MAXDIR (160) and added a check, return -1 if its too long.
The file selector only allowed 80 chars to be typed into the directory 
entry.

Made the file selector check that the path is less then FILE_MIXDIR, if 
you try and enter a path thats longer it will tell you that the path is 
too long, before it was writing into other memory and crashing.



git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@12318 954f8c5b-7b00-dc11-b283-0030488c597c