==Curve Editing== Curve select lasso and border didnt work with hide handles enabled.
Hiding handles was using the scene->selectmode variable, but looking at this now, I see there are places where selectmode is compared with 
SCE_SELECT_VERTEX, EDGE, FACE etc. as well as being.
Better to use a global flag, same as G_DRAWFACES.
This is not good since anyone using this in the few days will have an invalid selectmode variable. simple way to fix is to change modes, so do_versions shouldnt be needed.



git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@12391 954f8c5b-7b00-dc11-b283-0030488c597c