== Armature - Ghosting Feature (Cessen request) ==
Added a new option for Armature Ghosting: Only draw ghosts for selected bones. This is activated by toggling the "Sel" button beside the GStep: field. 

Note: this does not give any speed increases, as the whole pose must be recalculated for each ghost. In fact, it might even cause minor performance decreases, due to the need to tag and un-tag bones before/after drawing the set of ghosts.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@12829 954f8c5b-7b00-dc11-b283-0030488c597c